Properties + Border update
1) Single-argument setters for properties have been removed. The setter now has two arguments (the value and the value source), but the latter has a default value.
2) The Border element now has a working border width. It is displayed and taken into account when calculating layout.
3) Most fields of type size_t have been changed to uint32_t.
4) The line drawing method has been changed. Lines will now draw faster.

+++ Sgl/Base/Size.h
@@ -4,8 +4,8 @@ namespace Sgl
 {
 	struct Size
 	{
-		size_t Width;
-		size_t Height;
+		uint32_t Width;
+		uint32_t Height;
 	};
 
 	struct FSize
